backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1 
dj-database-url
psycopg2-binary 